Move protos => cli/server/
