+++ cli/server/protos/containers/v1/containers.proto
@@ -17,7 +17,7 @@ syntax = "proto3";
 
 package com.docker.api.protos.containers.v1;
 
-option go_package = "github.com/docker/compose-cli/protos/containers/v1;v1";
+option go_package = "github.com/docker/compose-cli/cli/server/protos/containers/v1;v1";
 
 service Containers {
 	rpc List(ListRequest) returns (ListResponse);
