+++ cli/server/proxy/containers_test.go
@@ -22,8 +22,8 @@ import (
 	"gotest.tools/v3/assert"
 
 	"github.com/docker/compose-cli/api/containers"
+	containersv1 "github.com/docker/compose-cli/cli/server/protos/containers/v1"
 	"github.com/docker/compose-cli/formatter"
-	containersv1 "github.com/docker/compose-cli/protos/containers/v1"
 )
 
 func TestGrpcContainerToContainerConfig(t *testing.T) {
