+++ cli/server/proxy/contexts.go
@@ -19,9 +19,9 @@ package proxy
 import (
 	"context"
 
+	contextsv1 "github.com/docker/compose-cli/cli/server/protos/contexts/v1"
 	"github.com/docker/compose-cli/config"
 	"github.com/docker/compose-cli/context/store"
-	contextsv1 "github.com/docker/compose-cli/protos/contexts/v1"
 )
 
 type contextsProxy struct {
