+++ cli/server/proxy/streams/io.go
@@ -19,7 +19,7 @@ package streams
 import (
 	"github.com/golang/protobuf/ptypes"
 
-	streamsv1 "github.com/docker/compose-cli/protos/streams/v1"
+	streamsv1 "github.com/docker/compose-cli/cli/server/protos/streams/v1"
 )
 
 // IO implements an io.ReadWriter that forwards everything to the stream
